requirements-dev.txt
# Include production dependencies
-r requirements.txt

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-flask==1.3.0

# Code quality
flake8==7.1.1
black==24.10.0
pylint==3.3.2
mypy==1.13.0

# Pre-commit hooks
pre-commit==4.0.1

# Security scanning
bandit==1.8.0
safety==3.2.11
